requirements.txt
Pillow==11.3.0
pydantic_settings==2.11.0
Requests==2.32.5
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
sqlalchemy_mixins==2.0.5
matplotlib==3.10.7
lxml==6.0.2
html5lib==1.1
beautifulsoup4==4.14.3